// frontend/apps/web/src/domains/api/generated/model/suggestionKind.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Applark
 * OpenAPI spec version: 0.1.0
 */

export type SuggestionKind = typeof SuggestionKind[keyof typeof SuggestionKind];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SuggestionKind = {
  emphasize: 'emphasize',
  rephrase: 'rephrase',
  add_detail: 'add_detail',
  deprioritize: 'deprioritize',
} as const;